requirements.txt
asgiref==3.8.1
Django==4.2.18
django-environ==0.12.0
django-role-permissions==3.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
mysqlclient==2.2.7
